requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.9
django-ckeditor==6.1.0
django-debug-toolbar==3.2.2
django-js-asset==1.2.2
Pillow==8.4.0
psycopg2==2.9.2
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
Unidecode==1.3.2
